# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Blog Writer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface Client {
  id: number;
  name: string;
  website: string;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  targetAudience?: string | null;
  /** @nullable */
  toneOfVoice?: string | null;
  searchConsoleConnected?: boolean;
  createdAt: string;
  updatedAt?: string;
}

export interface CreateClientBody {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  website: string;
  industry?: string;
  targetAudience?: string;
  toneOfVoice?: string;
}

export type BlogPostStatus =
  (typeof BlogPostStatus)[keyof typeof BlogPostStatus];

export const BlogPostStatus = {
  draft: "draft",
  published: "published",
  archived: "archived",
} as const;

export interface BlogPost {
  id: number;
  clientId: number;
  title: string;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  /** @nullable */
  focusKeyword?: string | null;
  status: BlogPostStatus;
  /** @nullable */
  wordCount?: number | null;
  /** @nullable */
  seoScore?: number | null;
  createdAt: string;
  updatedAt?: string;
}

export type CreateBlogBodyStatus =
  (typeof CreateBlogBodyStatus)[keyof typeof CreateBlogBodyStatus];

export const CreateBlogBodyStatus = {
  draft: "draft",
  published: "published",
  archived: "archived",
} as const;

export interface CreateBlogBody {
  /** @minLength 1 */
  title: string;
  content?: string;
  metaDescription?: string;
  focusKeyword?: string;
  status?: CreateBlogBodyStatus;
}

export type UpdateBlogBodyStatus =
  (typeof UpdateBlogBodyStatus)[keyof typeof UpdateBlogBodyStatus];

export const UpdateBlogBodyStatus = {
  draft: "draft",
  published: "published",
  archived: "archived",
} as const;

export interface UpdateBlogBody {
  title?: string;
  content?: string;
  metaDescription?: string;
  focusKeyword?: string;
  status?: UpdateBlogBodyStatus;
}

export interface GenerateBlogBody {
  /** @minLength 1 */
  topic: string;
  focusKeyword?: string;
  targetWordCount?: number;
  additionalInstructions?: string;
}

export interface ConnectSearchConsoleBody {
  /** @minLength 1 */
  siteUrl: string;
  accessToken?: string;
}

export interface SearchConsoleData {
  connected: boolean;
  siteUrl: string;
}

export interface KeywordData {
  keyword: string;
  clicks: number;
  impressions: number;
  ctr: number;
  position: number;
}

/**
 * Opaque session token — Bearer <sid>.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};