# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Blog Writer API
 * OpenAPI spec version: 0.1.0
 */
import type { BlogPostStatus } from "./blogPostStatus";

export interface BlogPost {
  id: number;
  clientId: number;
  title: string;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  /** @nullable */
  focusKeyword?: string | null;
  status: BlogPostStatus;
  /** @nullable */
  wordCount?: number | null;
  /** @nullable */
  seoScore?: number | null;
  createdAt: Date;
  updatedAt?: Date;
}