# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Blog Writer API
 * OpenAPI spec version: 0.1.0
 */

export interface Client {
  id: number;
  name: string;
  website: string;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  targetAudience?: string | null;
  /** @nullable */
  toneOfVoice?: string | null;
  searchConsoleConnected?: boolean;
  createdAt: Date;
  updatedAt?: Date;
}