# lib/api-zod/src/generated/types/connectSearchConsoleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Blog Writer API
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectSearchConsoleBody {
  /** @minLength 1 */
  siteUrl: string;
  accessToken?: string;
}